+++ bin/lighthouse.dart
@@ -2,7 +2,7 @@ import 'package:args/command_runner.dart';
 import 'package:lighthouse/src/commands/assets/add.dart';
 import 'package:lighthouse/src/commands/assets/make.dart';
 import 'package:lighthouse/src/commands/tr/make.dart';
-import 'package:lighthouse/src/exceptions/tr/tr_exception.dart';
+import 'package:lighthouse/src/exceptions/tr_exception.dart';
 
 Future<void> main(List<String> args) async {
   // await AssetsMakeCommand().run();
@@ -15,20 +15,20 @@ Future<void> main(List<String> args) async {
     // runner.addCommand(Init());
 
     /// * generate TR class
-    runner.addCommand(TR());
+    runner.addCommand(TRMake());
 
     /// * Assets
     /// * generate assets
     runner.addCommand(AssetsMakeCommand());
-        /// * add assets
+
+    /// * add assets
 
     runner.addCommand(AssetsAddCommand());
 
     await runner.run(args);
   } on TrException catch (e) {
-    print('LH :: Can\'t Generate TR File !');
-    print('LH ::' + e.toString());
-  } catch (e,st) {
+    print('💡❌:: Can\'t Generate TR File !\n$e');
+  } catch (e, st) {
     print(st);
     print(e.toString());
   }