+++ lib/src/exceptions/tr/tr_exception.dart
@@ -1,1 +1,0 @@
-class TrException implements Exception {}