+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 0.0.5
 
-- `lh tr:make` now adds nations default values to the `Tr` class (#8)
+- enhancements 🎁 ::
+  - `lh tr:make`
+    - adds nations default values to the `Tr` class (#8)
+    - validates all errors then report all failures
+    - will validates nested objects (#10)
 - NEW 🔥::
-  - `lh assets:add` command
--
+  - `lh assets:add` to add assets to `pubspec.yaml`